Refactor: Extract AtMemberMappable.
