+++ src/main/java/net/fabricmc/tinyremapper/extension/mixin/soft/annotation/injection/AtAnnotationVisitor.java
@@ -19,19 +19,15 @@
 package net.fabricmc.tinyremapper.extension.mixin.soft.annotation.injection;
 
 import java.util.Objects;
-import java.util.Optional;
 
 import org.objectweb.asm.AnnotationVisitor;
 import org.objectweb.asm.tree.AnnotationNode;
 
-import net.fabricmc.tinyremapper.api.TrMember;
 import net.fabricmc.tinyremapper.extension.mixin.common.IMappable;
-import net.fabricmc.tinyremapper.extension.mixin.common.ResolveUtility;
 import net.fabricmc.tinyremapper.extension.mixin.common.data.Annotation;
 import net.fabricmc.tinyremapper.extension.mixin.common.data.AnnotationElement;
 import net.fabricmc.tinyremapper.extension.mixin.common.data.CommonData;
 import net.fabricmc.tinyremapper.extension.mixin.common.data.Constant;
-import net.fabricmc.tinyremapper.extension.mixin.common.data.Message;
 import net.fabricmc.tinyremapper.extension.mixin.soft.data.MemberInfo;
 
 /**
@@ -87,7 +83,7 @@ class AtAnnotationVisitor extends AnnotationNode {
 					if (this.value.equals("NEW")) {
 						value = new AtConstructorMappable(data, info).result().toString();
 					} else {
-						value = new AtMethodMappable(data, info).result().toString();
+						value = new AtMemberMappable(data, info).result().toString();
 					}
 				}
 			}
@@ -119,36 +115,6 @@ class AtAnnotationVisitor extends AnnotationNode {
 			}
 
 			return av;
-		}
-	}
-
-	private static class AtMethodMappable implements IMappable<MemberInfo> {
-		private final CommonData data;
-		private final MemberInfo info;
-
-		AtMethodMappable(CommonData data, MemberInfo info) {
-			this.data = Objects.requireNonNull(data);
-			this.info = Objects.requireNonNull(info);
-		}
-
-		@Override
-		public MemberInfo result() {
-			if (!info.isFullyQualified()) {
-				data.getLogger().warn(Message.NOT_FULLY_QUALIFIED, info);
-				return info;
-			}
-
-			Optional<TrMember> resolved = data.resolver.resolveMember(info.getOwner(), info.getName(), info.getDesc(), ResolveUtility.FLAG_UNIQUE | ResolveUtility.FLAG_RECURSIVE);
-
-			if (resolved.isPresent()) {
-				String newOwner = data.mapper.asTrRemapper().map(info.getOwner());
-				String newName = data.mapper.mapName(resolved.get());
-				String newDesc = data.mapper.mapDesc(resolved.get());
-
-				return new MemberInfo(newOwner, newName, info.getQuantifier(), newDesc);
-			} else {
-				return info;
-			}
 		}
 	}
